requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
darksky-weather==1.6.5
idna==2.8
multidict==4.7.4
pytz==2019.1
requests==2.21.0
urllib3==1.24.3
wincertstore==0.2
yarl==1.4.2
